A small program to keep track of customers and there services
A family business project that probably took way too long to implement,
but at least I learned at least 50 different ways to use databases in
.NET. The UI could certainly use some work, and almost nothing is
commented, but the code is fairly straight-forward anyway and it worked
for the family member in the company that would be using it. This is
certainly a project I look forward to expanding in the future.